Change output name of the static library on Windows
Since both static and import libraries have the same file extension
on Windows (.lib), except when building with MinGW or Cygwin, we have
to somewhoe distinguish between them. A common practice is to append
a suffix to the filename, such as "_static"., so the resulting binary
will be "subhook_static.lib".