Modify x64 support to use RET rather than JMP
Unforunately, 64-bit direct jumps don't work in AMD64. In order to
avoid this issue, I am pushing the address we wish to jump to on the
stack and calling RET.

Due to the same issues as JMP, using the trampoline is unlikely to be
safe in many situations. Therefore, I would recommend only using
ScopedRemove, or manually removing the hook prior to calling the
original function, instead of using a the trampoline.