Replace test_exe with asm_test_exe
test_exe was written in C and depended a lot on the compiler and
compiler/linker flags being used, which made it unpredictable sometimes.

On the other hand, asm_test_exe is written  assembly, so it generates the
same code each time, which is easier to rely on for testing purposes.