unix: Fix unprotecting memory across a page boundary
Previously, we'd round the address to unprotect to the nearest page but
keep the length the same. This meant that if we were hooking a function
that started just before the end of a page, it would need both pages to
be unprotected but we'd only unprotect the first one.